requirements.txt
discord.py
python-dotenv
requests
pymongo[srv]
asyncio